[Leveler] Add default bg settings
When a user has deleted the images named "default" for the profile, rank, and levelup bgs, the _create_user function had those hardcoded in as the default images to use for a new user. 
This change adds [p]lvladmin bg defaultrankbg, [p]lvladmin bg defaultprofilebg, and [p]lvladmin bg defaultlevelbg so that bot owners can set the default backgrounds to use for new user creation or new entries (if an existing user calls "rank" for themselves for the first time, for example).
There is also new logic for making sure that the bot owner doesn't delete every profile, rank, and levelup bg in the database without having any to choose from - there must always be >= 1 item in the profile, rank, and levelup bg lists.